chore: rename and shuffle
