+++ blocks/header/header.js
@@ -1,4 +1,4 @@
-import { readBlockConfig, decorateIcons } from '../../scripts/lib-helix.js';
+import { readBlockConfig, decorateIcons } from '../../scripts/block-utils.js';
 
 /**
  * collapses all open nav sections
