+++ scripts/delayed.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { sampleRUM } from './lib-helix.js';
+import { sampleRUM } from './block-utils.js';
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
